feat(domain): add  ModbusConfig dataclass

+++ domain/modbus_config.py
@@ -5,4 +5,4 @@ from dataclasses import dataclass
 class ModbusConfig:
     host: str
     port: int
-    poll_interval_s: int
+    poll_interval_s: float
